Draft: toggle between grid and list on Tab Switcher screen
