+++ DuckDuckGo/Assets.xcassets/TabsToggleGrid.imageset/tabsToggleGrid.pdf
@@ -1,0 +1,131 @@
+%PDF-1.7
+
+1 0 obj
+  << >>
+endobj
+
+2 0 obj
+  << /Length 3 0 R >>
+stream
+/DeviceRGB CS
+/DeviceRGB cs
+q
+1.000000 0.000000 -0.000000 1.000000 0.000000 0.000000 cm
+0.200000 0.200000 0.200000 scn
+1.000000 16.000000 m
+0.447715 16.000000 0.000000 15.552285 0.000000 15.000000 c
+0.000000 10.000000 l
+0.000000 9.447716 0.447715 9.000000 1.000000 9.000000 c
+6.000000 9.000000 l
+6.552285 9.000000 7.000000 9.447716 7.000000 10.000000 c
+7.000000 15.000000 l
+7.000000 15.552284 6.552285 16.000000 6.000000 16.000000 c
+1.000000 16.000000 l
+h
+1.000000 7.000000 m
+0.447715 7.000000 0.000000 6.552285 0.000000 6.000000 c
+0.000000 1.000000 l
+0.000000 0.447716 0.447715 0.000000 1.000000 0.000000 c
+6.000000 0.000000 l
+6.552285 0.000000 7.000000 0.447715 7.000000 1.000000 c
+7.000000 6.000000 l
+7.000000 6.552284 6.552285 7.000000 6.000000 7.000000 c
+1.000000 7.000000 l
+h
+9.000000 15.000000 m
+9.000000 15.552285 9.447716 16.000000 10.000000 16.000000 c
+15.000000 16.000000 l
+15.552285 16.000000 16.000000 15.552284 16.000000 15.000000 c
+16.000000 10.000000 l
+16.000000 9.447716 15.552284 9.000000 15.000000 9.000000 c
+10.000000 9.000000 l
+9.447715 9.000000 9.000000 9.447716 9.000000 10.000000 c
+9.000000 15.000000 l
+h
+10.000000 7.000000 m
+9.447716 7.000000 9.000000 6.552285 9.000000 6.000000 c
+9.000000 1.000000 l
+9.000000 0.447716 9.447715 0.000000 10.000000 0.000000 c
+15.000000 0.000000 l
+15.552284 0.000000 16.000000 0.447715 16.000000 1.000000 c
+16.000000 6.000000 l
+16.000000 6.552284 15.552285 7.000000 15.000000 7.000000 c
+10.000000 7.000000 l
+h
+f*
+n
+Q
+q
+1.000000 0.000000 -0.000000 1.000000 27.000000 0.000000 cm
+0.600000 0.600000 0.600000 scn
+1.000000 16.000000 m
+0.447715 16.000000 0.000000 15.552285 0.000000 15.000000 c
+0.000000 10.000000 l
+0.000000 9.447716 0.447715 9.000000 1.000000 9.000000 c
+15.000000 9.000000 l
+15.552285 9.000000 16.000000 9.447716 16.000000 10.000000 c
+16.000000 15.000000 l
+16.000000 15.552284 15.552285 16.000000 15.000000 16.000000 c
+1.000000 16.000000 l
+h
+1.000000 7.000000 m
+0.447715 7.000000 0.000000 6.552285 0.000000 6.000000 c
+0.000000 1.000000 l
+0.000000 0.447716 0.447715 0.000000 1.000000 0.000000 c
+15.000000 0.000000 l
+15.552285 0.000000 16.000000 0.447715 16.000000 1.000000 c
+16.000000 6.000000 l
+16.000000 6.552284 15.552285 7.000000 15.000000 7.000000 c
+1.000000 7.000000 l
+h
+f*
+n
+Q
+
+endstream
+endobj
+
+3 0 obj
+  2251
+endobj
+
+4 0 obj
+  << /Annots []
+     /Type /Page
+     /MediaBox [ 0.000000 0.000000 43.000000 16.000000 ]
+     /Resources 1 0 R
+     /Contents 2 0 R
+     /Parent 5 0 R
+  >>
+endobj
+
+5 0 obj
+  << /Kids [ 4 0 R ]
+     /Count 1
+     /Type /Pages
+  >>
+endobj
+
+6 0 obj
+  << /Type /Catalog
+     /Pages 5 0 R
+  >>
+endobj
+
+xref
+0 7
+0000000000 65535 f
+0000000010 00000 n
+0000000034 00000 n
+0000002341 00000 n
+0000002364 00000 n
+0000002537 00000 n
+0000002611 00000 n
+trailer
+<< /ID [ (some) (id) ]
+   /Root 6 0 R
+   /Size 7
+>>
+startxref
+2670
+%%EOF
