+++ DuckDuckGo/TabSwitcherViewController.swift
@@ -37,7 +37,7 @@ class TabSwitcherViewController: UIViewController {
     @IBOutlet weak var collectionView: UICollectionView!
     @IBOutlet weak var toolbar: UIToolbar!
     
-    @IBOutlet weak var settingsButton: UIButton!
+    @IBOutlet weak var displayModeButton: UIButton!
     @IBOutlet weak var bookmarkAllButton: UIButton!
     
     @IBOutlet weak var fireButton: UIBarButtonItem!
@@ -55,6 +55,7 @@ class TabSwitcherViewController: UIViewController {
     
     var currentSelection: Int?
     
+    private var isGridEnabled = true
     private var isProcessingUpdates = false
 
     override func viewDidLoad() {
@@ -66,10 +67,28 @@ class TabSwitcherViewController: UIViewController {
         becomeFirstResponder()
     }
     
-    func setupBackgroundView() {
+    private func setupBackgroundView() {
         let view = UIView(frame: collectionView.frame)
         view.addGestureRecognizer(UITapGestureRecognizer(target: self, action: #selector(handleTap(gesture:))))
         collectionView.backgroundView = view
+    }
+    
+    private func refreshDisplayModeButton(theme: Theme = ThemeManager.shared.currentTheme) {
+        switch theme.currentImageSet {
+        case .dark:
+            // Reverse colors (selection)
+            if isGridEnabled {
+                displayModeButton.setImage(UIImage(named: "TabsToggleList"), for: .normal)
+            } else {
+                displayModeButton.setImage(UIImage(named: "TabsToggleGrid"), for: .normal)
+            }
+        case .light:
+            if isGridEnabled {
+                displayModeButton.setImage(UIImage(named: "TabsToggleGrid"), for: .normal)
+            } else {
+                displayModeButton.setImage(UIImage(named: "TabsToggleList"), for: .normal)
+            }
+        }
     }
 
     override func viewWillAppear(_ animated: Bool) {
@@ -197,9 +216,11 @@ class TabSwitcherViewController: UIViewController {
         present(alert, animated: true, completion: nil)
     }
     
-    @IBAction func onSettingsPressed(_ sender: UIButton) {
-        // Segue performed from storyboard
-        Pixel.fire(pixel: .settingsOpenedFromTabsSwitcher)
+    @IBAction func onDisplayModeButtonPressed(_ sender: UIButton) {
+        isGridEnabled = !isGridEnabled
+        
+        refreshDisplayModeButton()
+        collectionView.reloadData()
     }
 
     @IBAction func onAddPressed(_ sender: UIBarButtonItem) {
@@ -282,8 +303,10 @@ extension TabSwitcherViewController: UICollectionViewDataSource {
     }
 
     public func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: TabViewCell.reuseIdentifier, for: indexPath) as? TabViewCell else {
-            fatalError("Failed to dequeue cell \(TabViewCell.reuseIdentifier) as TablViewCell")
+        
+        let cellIdentifier = isGridEnabled ? TabViewGridCell.reuseIdentifier : TabViewListCell.reuseIdentifier
+        guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: cellIdentifier, for: indexPath) as? TabViewCell else {
+            fatalError("Failed to dequeue cell \(cellIdentifier) as TabViewCell")
         }
         cell.delegate = self
         cell.isDeleting = false
@@ -343,7 +366,7 @@ extension TabSwitcherViewController: UICollectionViewDelegateFlowLayout {
         
         // Calculate height based on the view size
         let contentAspectRatio = collectionView.bounds.width / collectionView.bounds.height
-        let heightToFit = (columnWidth / contentAspectRatio) + TabViewCell.Constants.cellHeaderHeight
+        let heightToFit = (columnWidth / contentAspectRatio) + TabViewGridCell.Constants.cellHeaderHeight
         
         // Try to display at least `preferredMinNumberOfRows`
         let preferredMaxHeight = collectionView.bounds.height / Constants.preferredMinNumberOfRows
@@ -357,10 +380,17 @@ extension TabSwitcherViewController: UICollectionViewDelegateFlowLayout {
                         layout collectionViewLayout: UICollectionViewLayout,
                         sizeForItemAt indexPath: IndexPath) -> CGSize {
         
-        let columnWidth = calculateColumnWidth(minimumColumnWidth: 150, maxColumns: 4)
-        let rowHeight = calculateRowHeight(columnWidth: columnWidth)
-        return CGSize(width: floor(columnWidth),
-                      height: floor(rowHeight))
+        if isGridEnabled {
+            let columnWidth = calculateColumnWidth(minimumColumnWidth: 150, maxColumns: 4)
+            let rowHeight = calculateRowHeight(columnWidth: columnWidth)
+            return CGSize(width: floor(columnWidth),
+                          height: floor(rowHeight))
+        } else {
+            let layout = collectionView.collectionViewLayout as? UICollectionViewFlowLayout
+            let spacing = layout?.sectionInset.left ?? 0.0
+            
+            return CGSize(width: collectionView.bounds.size.width - 2 * spacing, height: 70)
+        }
     }
     
 }
@@ -382,8 +412,9 @@ extension TabSwitcherViewController: Themable {
     func decorate(with theme: Theme) {
         view.backgroundColor = theme.backgroundColor
         
+        refreshDisplayModeButton(theme: theme)
+        
         titleView.textColor = theme.barTintColor
-        settingsButton.tintColor = theme.barTintColor
         bookmarkAllButton.tintColor = theme.barTintColor
         
         toolbar.barTintColor = theme.barBackgroundColor
